Middleware implementation

+++ routes/web.php
@@ -31,31 +31,31 @@ Route::get('/', [ListingController::class,'index'] );
 
 
 //Show create form
-Route::get('/listings/create', [ListingController::class, 'create'] );
+Route::get('/listings/create', [ListingController::class, 'create'] )->middleware('auth');
 
 // Store Listing Data
-Route::post('/listings', [ListingController::class, 'store'] );
+Route::post('/listings', [ListingController::class, 'store'] )->middleware('auth');
 
 //Show Edit Form  // Bracket er moddar variable muloto edit function er variable
-Route::get('/listings/{my_listing}/edit', [ListingController::class,'edit'] );
+Route::get('/listings/{my_listing}/edit', [ListingController::class,'edit'] )->middleware('auth');
 
 //Edit Submit to Update
-Route::put('/listings/{listing}', [ListingController::class,'update'] );
+Route::put('/listings/{listing}', [ListingController::class,'update'] )->middleware('auth');
 
 //Delete lisitng
-Route::delete('/listings/{listing}', [ListingController::class,'destroy'] );
+Route::delete('/listings/{listing}', [ListingController::class,'destroy'] )->middleware('auth');
 
 // Show Register/Create Form
-Route::get('/register', [UserController::class,'create'] );
+Route::get('/register', [UserController::class,'create'] )->middleware('guest');;
 
 // Create New User
 Route::post('/users', [UserController::class,'store'] );
 
 // Log User Out
-Route::post('/logout', [UserController::class, 'logout'] );
+Route::post('/logout', [UserController::class, 'logout'] )->middleware('auth');
 
 // Show login Form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login');
 
 //Log In User
 Route::post('/users/authenticate', [UserController::class,'authenticate'] );
